apps/requirements.txt
confluent-kafka>=1.6.1
cassandra-driver>=3.25.0
python-dateutil>=2.8.1
